Initial KDS service
The Key Distribution Service is used to register keys for services and
distribute tickets to contact othe services.

The KDS is used to digitally sign and optionally encrypt messages sent over the
message queue by the rpc modules.

It implements the service described in this document:
https://wiki.openstack.org/wiki/MessageSecurity#A_Key_Distribution_Server_in_Keystone

DocImpact
SecurityImpact
blueprint key-distribution-server
Change-Id: Ib47aca8f72623a07ff18f23d46d0af520e463fc9
Co-Authored-By: Jamie Lennox <jamielennox@redhat.com>