Remove host from per notification options
In oslo.messaging publisher_id is expected to be set in a one of.
Specifying host allows you to set the sender so that messages come from
'identity.host' however this is something you are far more likely to
want to set up per instance that modify when actually emitting a
notification. As there is nothing so far using this parameter
remove it.

Change-Id: Ie70c313bcebd67cc8c6a3d25eba95405089bd44a
blueprint: oslo-messaging