Make Pecan the root routing framework
Replace the core of keystone routing with Pecan. All the original
controllers are still used and handed off if there is not a valid pecan
route (which is only versioning now).

There are a number of changes here together because they don't
always work separately. Things like moving exception handling to a hook
can work individually but is difficult to mix and match with the
existing infrastructure.

In this patch we start to separate the v2 and v3 APIs into individual
folders that can be continued later.

Blueprint: keystone-pecan
Change-Id: I0172ced01bb4017d184f1d2091dba7444e59df8b