Don't use super for log handler
RHOS 5 supports RHEL 6 which has python 2.6 at which point log handlers
were still old style classes and so super didn't work.

Whilst this isn't a problem on RHEL7 with python 2.7 it also doesn't
hurt and so it's easier to fix it for both than ship a patch for only
the RHEL 6 version.

Resolves rhbz: 1264515

Change-Id: I42539494f35ea00f5c4dc9adaeb9db805856803b