Ignore unknown arguments to fetch_token
fetch_token is an implementation of a method from auth_token middleware.
As we need to add new features to auth_token middleware we may need to
add new information to the fetch_token method which keystone won't
understand immediately.

We define on the auth_token middleware side that an implementation
should ignore any unknown keyword arguments passed to fetch_token to
allow this to change.

Change-Id: I0cd4c414e496b99b01b3f7f19027f4cf92d17d71